自訂 is more appropriate
自訂 is more appropriate. It means custom or customized.   自定義 means self-defining and it is way too heavy.  It is not "self defining". It just means a custom message. In fact, it could have been 我們把網頁標題設定為包含點擊次數的訊息。

The word "custom" is not even needed, because we are saying "we set it to a message that includes the number of clicks", and the meaning of "custom" is very apparent.  We could use the word 自訂 if we want to stick to the original English sentence as close as possible.